Profile page + backend Index page refactoring

+++ src/main/java/com/tsystems/ecare/app/dao/ContractDao.java
@@ -2,10 +2,8 @@ package com.tsystems.ecare.app.dao;
 
 import com.tsystems.ecare.app.model.Contract;
 
-/**
- * @author Andrei Makarevich
- */
-public interface ContractDao extends GenericDao<Contract, Integer> {
+
+public interface ContractDao extends GenericDao<Contract, Long> {
 
     public Contract findByNumber(String number);
 }
